requirements.txt
pandas
matplotlib
langchain
langchain-openai
python-dotenv
gradio
ipykernel
